requirements.txt
black==22.3.0
click==8.1.3
mypy-extensions==0.4.3
numpy==1.22.4
pandas==1.4.2
pathspec==0.9.0
Pillow==9.1.1
platformdirs==2.5.2
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
tomli==2.0.1
tqdm==4.64.0
typing_extensions==4.2.0
